@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module MyApiClient
-  VERSION = '0.22.0'
+  VERSION = '0.23.0'
 end
