@@ -1,46 +1,62 @@
 # frozen_string_literal: true
 
 RSpec.describe MyApiClient::Error do
-  let(:instance) { described_class.new(params) }
-  let(:params) do
-    instance_double(
-      MyApiClient::Params::Params,
-      inspect: '"#<MyApiClient::Params::Params#inspect>"',
-      metadata: { request: 'params', response: 'params' }
-    )
-  end
+  context 'when initialized with params and error message' do
+    let(:instance) { described_class.new(params, 'error message') }
 
-  describe '#params' do
-    it 'returns a params instance' do
-      expect(instance.params).to eq params
+    let(:params) do
+      instance_double(
+        MyApiClient::Params::Params,
+        inspect: '"#<MyApiClient::Params::Params#inspect>"',
+        metadata: { request: 'params', response: 'params' }
+      )
+    end
+
+    describe '#params' do
+      it 'returns a params instance' do
+        expect(instance.params).to eq params
+      end
+    end
+
+    describe '#metadata' do
+      it 'delegates processing to params#metadata' do
+        expect(instance.metadata).to eq(request: 'params', response: 'params')
+      end
+    end
+
+    describe '#inspect' do
+      it 'returns contents as string for to be readable for human' do
+        expect(instance.inspect)
+          .to eq '{:error=>"#<MyApiClient::Error: error message>", ' \
+                 ':params=>"#<MyApiClient::Params::Params#inspect>"}'
+      end
+    end
+
+    describe '#message' do
+      it { expect(instance.message).to eq 'error message' }
     end
   end
 
-  describe '#metadata' do
-    it 'delegates processing to params#metadata' do
-      expect(instance.metadata).to eq(request: 'params', response: 'params')
-    end
-  end
+  context 'when initialized with no arguments (for RSpec)' do
+    let(:instance) { described_class.new }
 
-  describe '#inspect' do
-    it 'returns contents as string for to be readable for human' do
-      expect(instance.inspect)
-        .to eq '{:error=>"#<MyApiClient::Error: MyApiClient::Error>", ' \
-               ':params=>"#<MyApiClient::Params::Params#inspect>"}'
-    end
-  end
-
-  describe '#message' do
-    subject { instance.message }
-
-    context 'when no error message is given' do
-      it { is_expected.to eq 'MyApiClient::Error' }
+    describe '#params' do
+      it { expect(instance.params).to be_nil }
     end
 
-    context 'when error message is given' do
-      let(:instance) { described_class.new(params, 'error message') }
+    describe '#metadata' do
+      it { expect(instance.metadata).to be_nil }
+    end
 
-      it { is_expected.to eq 'error message' }
+    describe '#inspect' do
+      it 'returns contents as string for to be readable for human' do
+        expect(instance.inspect)
+          .to eq '{:error=>"#<MyApiClient::Error: MyApiClient::Error>", :params=>nil}'
+      end
+    end
+
+    describe '#message' do
+      it { expect(instance.message).to eq 'MyApiClient::Error' }
     end
   end
 end
